Update dto and entities

+++ src/main/java/app/servlets/fitness/dto/SubscriptionDto.java
@@ -4,12 +4,14 @@ import app.servlets.fitness.entities.enums.SubscriptionCategory;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class SubscriptionDto {
     private SubscriptionCategory subscriptionCategory;
     private String subscriptionName;
-    private int subscriptionPrice;
+    private BigDecimal subscriptionPrice;
     private int subscriptionDaysNumber;
     private int numberGuestVisitDays;
     private int numberSubscriptionStopDays;
